Reduce docker image space (#21)
# Motivation
This reduces the space needed for the dev setup by avoiding having
separate virtual environments for every compose service, and instead
storing one virtual environment in the docker image itself.
Additionally, this approach still allows usage of `uv` outside the
docker container itself, so nothing is lost with this approach.

# Possible Problems
This no longer puts the virtual environment on `PATH`, so usage of `uv
run` will be necessary.